Extracting tasks out to separate files grouped logically
