+++ addon/index.js
@@ -2,11 +2,8 @@ export {
   runTask,
   scheduleTask,
   throttleTask,
-  pollTask,
-  pollTaskFor,
-  setShouldPoll,
   cancelTask,
   cancelDebounce,
-  cancelPoll,
 } from './utils/tasks';
+export { pollTask, pollTaskFor, setShouldPoll, cancelPoll } from './poll-task';
 export { runDisposables } from './utils/disposable';
